+++ client_src/js/constants/history.constants.js
@@ -1,0 +1,8 @@
+module.exports = {
+  HISTORY_CONSTANTS: {
+    currentPage: "page=",
+    perPage: "limit=",
+    sort: "sort="
+  },
+  SORT_CONSTANTS: {}
+};